feat: cycle through carousel component images
clicking on the left or right button of the carousel component will
cycle to the next image and display it. If the count of the cycle is out
of bounds of the array of images, it will cycle to the nearest starting
point of the image array (start or end of the array).